Moved repeated functions to a single file in misc folder.
