+++ Tools/runStats
@@ -5,27 +5,19 @@ import sys
 import getopt
 import codecs
 import re
-import urllib2
-import logging
 import os
 from fontTools import subset, compare, treeshake
 from fontTools.misc.captureOutput import captureOutput
-import unicodedata
+from fontTools.misc.inputParsing import getCharListFromInput
+from fontTools.misc.util import makeArgs, makeOutputFileName
 from fontTools.ttLib import TTFont
-from bs4 import BeautifulSoup
 from os import rename, listdir
-from urllib2 import urlopen
 from cStringIO import StringIO
 
 def usage():
     print("usage: runStats inputFile")
     sys.exit(1)        
 
-def makeArgs(*params):
-    args = []
-    for param in params:
-        args.append(param)
-    return args
 
 def treeShake(fontFile):
     args = makeArgs(fontFile)
@@ -34,58 +26,6 @@ def treeShake(fontFile):
 
     reducedFontFile = 'Reduced'+fontFile
     return reducedFontFile
-
-def makeOutputFileName(input, extension):
-    dirName, fileName = os.path.split(input)
-    fileName, ext = os.path.splitext(fileName)
-    output = os.path.join(dirName, fileName + extension)
-    n = 1
-    while os.path.exists(output):
-        output = os.path.join(dirName, fileName + "#" + repr(n) + extension)
-        n = n + 1
-    return output
-
-def readFile(input, encoding):
-    try:
-        with codecs.open(input) as file:
-            data=file.read()
-    except:
-        raise IOError("Couldn't open file "+input)
-
-    try:
-        data = data.decode(encoding)
-    except:
-        raise ValueError("Different encoding or wrong code provided: "+encoding)
-    return data
-
-def readPage(input):
-    #some sites block common non-browser user agent strings, better use a regular one
-    hdr = {'User-Agent':'Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US; rv:1.9.0.7) Gecko/2009021910 Firefox/3.0.7'}
-    request = urllib2.Request(input, headers = hdr)
-    try:
-        page = urllib2.urlopen(request)
-    except:
-        raise ValueError("Couldn't load file or URL "+input)
-    return page
-
-def readInput(input, encoding):
-    try:
-        data = readFile(input, encoding)
-    except IOError:
-        data = readPage(input)
-
-    soup = BeautifulSoup(data)
-    texts = soup.findAll(text=True)
-    [s.extract() for s in soup(['style', 'script', '[document]', 'head', 'title'])]
-    string = soup.getText()
-
-    charList = list(set(string))
-
-    #reference to unicode categories http://www.sql-und-xml.de/unicode-database/#kategorien
-    non_printable = ['Zs', 'Zl', 'Zp', 'Cc']
-    charList = [x for x in charList if unicodedata.category(x) not in non_printable]
-
-    return charList
 
 def generateStats(stats):
 
@@ -166,7 +106,7 @@ def main(args):
     results = []
     for fontFile, inputFile, encoding in testSuite:
         tsFontFile = treeShake(fontFile) 
-        charList = readInput(inputFile, encoding)
+        charList = getCharListFromInput(inputFile, encoding)
         glyphs = "--text="+''.join(charList)
 
         filesBefore = listdir(".")
